Fixed all encryption refactor bugs and added in more changes to the changelog
