+++ library/src/main/res/drawable-v21/custom_background_white_back_black_edges_heavy_rounding.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- Button -->
+    <item>
+        <shape xmlns:android="http://schemas.android.com/apk/res/android"
+            android:shape="rectangle">
+
+            <gradient
+                android:angle="270"
+                android:endColor="@color/white"
+                android:startColor="@color/white" />
+
+            <stroke
+                android:width="1dp"
+                android:color="@color/black" />
+            <corners
+                android:bottomLeftRadius="16dp"
+                android:bottomRightRadius="16dp"
+                android:topLeftRadius="16dp"
+                android:topRightRadius="16dp" />
+
+        </shape>
+    </item>
+
+    <!-- Ripple -->
+    <item>
+        <ripple xmlns:android="http://schemas.android.com/apk/res/android"
+            android:color="?attr/colorControlHighlight" />
+    </item>
+
+</layer-list>
