+++ library/src/main/res/drawable/custom_background_white_back_black_edges_heavy_rounding.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <gradient
+        android:startColor="@color/white"
+        android:endColor="@color/white"
+        android:angle="270"/>
+
+    <stroke
+        android:width="1dp"
+        android:color="@color/black"
+        />
+    <corners
+        android:bottomRightRadius="16dp"
+        android:bottomLeftRadius="16dp"
+        android:topLeftRadius="16dp"
+        android:topRightRadius="16dp"/>
+
+</shape>
